refactor: move stories and tests
