+++ packages/components/skeleton/stories/SkeletonContainer.stories.tsx
@@ -1,8 +1,11 @@
 import React from 'react';
 
-import { SkeletonBodyText } from '../SkeletonBodyText/SkeletonBodyText';
-import { SkeletonDisplayText } from '../SkeletonDisplayText/SkeletonDisplayText';
-import { SkeletonContainer, SkeletonContainerProps } from './SkeletonContainer';
+import { SkeletonBodyText } from '../src/SkeletonBodyText/SkeletonBodyText';
+import { SkeletonDisplayText } from '../src/SkeletonDisplayText/SkeletonDisplayText';
+import {
+  SkeletonContainer,
+  SkeletonContainerProps,
+} from '../src/SkeletonContainer/SkeletonContainer';
 
 export default {
   title: 'Components/Skeleton/SkeletonContainer',
