+++ packages/components/skeleton/stories/SkeletonDisplayText.stories.tsx
@@ -1,10 +1,10 @@
 import React from 'react';
 
-import { SkeletonContainer } from '../SkeletonContainer/SkeletonContainer';
+import { SkeletonContainer } from '../src/SkeletonContainer/SkeletonContainer';
 import {
   SkeletonDisplayText,
   SkeletonDisplayTextProps,
-} from './SkeletonDisplayText';
+} from '../src/SkeletonDisplayText/SkeletonDisplayText';
 
 export default {
   title: 'Components/Skeleton/SkeletonDisplayText',
