+++ packages/components/skeleton/test/SkeletonBodyText.test.tsx
@@ -2,8 +2,8 @@ import React from 'react';
 import { render } from '@testing-library/react';
 import { axe } from '@/scripts/test/axeHelper';
 
-import { SkeletonContainer } from '../SkeletonContainer/SkeletonContainer';
-import { SkeletonBodyText } from './SkeletonBodyText';
+import { SkeletonContainer } from '../src/SkeletonContainer/SkeletonContainer';
+import { SkeletonBodyText } from '../src/SkeletonBodyText/SkeletonBodyText';
 
 it('has no a11y issues', async () => {
   const { container } = render(
