+++ packages/components/skeleton/test/SkeletonRow.test.tsx
@@ -3,7 +3,7 @@ import { render } from '@testing-library/react';
 import { Table, TableBody } from '@contentful/f36-table';
 import { axe } from '@/scripts/test/axeHelper';
 
-import { SkeletonRow } from './SkeletonRow';
+import { SkeletonRow } from '../src/SkeletonRow/SkeletonRow';
 
 it('renders the component', () => {
   const { container } = render(
